+++ app/app.php
@@ -1,24 +1,25 @@
 <?php
 
 /**
- *
- * @package Olympus
- * @author Achraf Chouk <achrafchouk@gmail.com>
- * @since 0.0.1
- *
+ * Autoload vendors and set all defined WordPress constants.
+ * 
+ * @category PHP
+ * @package  Olympus
+ * @author   Achraf Chouk <achrafchouk@gmail.com>
+ * @license  https://github.com/crewstyle/Olympus/blob/master/LICENSE MIT
+ * @link     https://github.com/crewstyle/Olympus
+ * @since    0.0.1
  */
 
 /**
  * Path to WordPress.
  */
-if (!defined('ABSPATH')) {
-    define('ABSPATH', WEBPATH.'cms'.S);
-}
+defined('ABSPATH') or define('ABSPATH', WEBPATH.'cms'.S);
 
 /**
  * Include composer autoloading.
  */
-$loader = require_once APPPATH.'autoload.php';
+$loader = include_once APPPATH.'autoload.php';
 
 /**
  * Load environment configuration.