+++ app/autoload.php
@@ -1,35 +1,34 @@
 <?php
 
 /**
- *
- * @package Olympus
- * @author Achraf Chouk <achrafchouk@gmail.com>
- * @since 0.0.1
- *
+ * Register the composer autoloader.
+ * 
+ * @category PHP
+ * @package  Olympus
+ * @author   Achraf Chouk <achrafchouk@gmail.com>
+ * @license  https://github.com/crewstyle/Olympus/blob/master/LICENSE MIT
+ * @link     https://github.com/crewstyle/Olympus
+ * @since    0.0.1
  */
 
 /**
- * Register rhe composer autoloader.
+ * Register the composer autoloader.
  */
 if (!file_exists($autoload = VENDORPATH.'autoload.php')) {
-    die('
-        <h1>Unable to find composer autoloader.</h1>
-        Please use: <code>curl -s http://getcomposer.org/installer | php</code> 
-        and <code>php composer.phar install</code>
-    ');
+    die('<h1>Unable to find composer autoloader.</h1> Please use: <code>curl -s http://getcomposer.org/installer | php</code> and <code>php composer.phar install</code>');
 }
 
-$loader = require $autoload;
+$loader = include $autoload;
 
 /**
  * Include the compiled class file.
  *
  * To dramatically increase your application's performance, you may use a
  * compiled class file which contains all of the classes commonly used
- * by a request. The Artisan "optimize" is used to create this file.
+ * by a request.
  */
 if (file_exists($compiled = APPPATH.'cache'.S.'compiled.php')) {
-    require $compiled;
+    include $compiled;
 }
 
 /**