+++ web/index.php
@@ -1,11 +1,14 @@
 <?php
 
 /**
- *
- * @package Olympus
- * @author Achraf Chouk <achrafchouk@gmail.com>
- * @since 0.0.1
- *
+ * Bootstrap all WordPress context.
+ * 
+ * @category PHP
+ * @package  Olympus
+ * @author   Achraf Chouk <achrafchouk@gmail.com>
+ * @license  https://github.com/crewstyle/Olympus/blob/master/LICENSE MIT
+ * @link     https://github.com/crewstyle/Olympus
+ * @since    0.0.1
  */
 
 $s = DIRECTORY_SEPARATOR;