+++ web/wp-config.php
@@ -1,11 +1,14 @@
 <?php
 
 /**
- *
- * @package Olympus
- * @author Achraf Chouk <achrafchouk@gmail.com>
- * @since 0.0.1
- *
+ * File used by WordPress to list all constant options.
+ * 
+ * @category PHP
+ * @package  Olympus
+ * @author   Achraf Chouk <achrafchouk@gmail.com>
+ * @license  https://github.com/crewstyle/Olympus/blob/master/LICENSE MIT
+ * @link     https://github.com/crewstyle/Olympus
+ * @since    0.0.1
  */
 
 $path = dirname(dirname(__FILE__));
@@ -17,7 +20,6 @@ $path = dirname(dirname(__FILE__));
 defined('S')           or define('S',           DIRECTORY_SEPARATOR);
 // Paths.
 defined('APPPATH')     or define('APPPATH',     $path.S.'app'.S);
-defined('LIBPATH')     or define('LIBPATH',     $path.S.'library'.S);
 defined('VENDORPATH')  or define('VENDORPATH',  $path.S.'vendor'.S);
 defined('WEBPATH')     or define('WEBPATH',     $path.S.'web'.S);
 // Contents folder